Split long, 1k+ line SyncUpDown into 3 files.
